+++ conf/urls.py
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.conf import settings
 from django.contrib import admin
-from django.urls import path, re_path, include
+from django.urls import path, include
 from infra.apps.common import views as common_views
 
 urlpatterns = [
